requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
PyMuPDF
chromadb
python-multipart